+++ pkg/skaffold/runner/verify.go
@@ -56,11 +56,13 @@ func (r *SkaffoldRunner) Verify(ctx context.Context, out io.Writer, artifacts []
 
 	out, ctx = output.WithEventContext(ctx, out, constants.Verify, constants.SubtaskIDNone)
 
-	output.Default.Fprintln(out, "Tags used in verification:")
+	if len(artifacts) > 0 {
+		output.Default.Fprintln(out, "Tags used in verification:")
 
-	for _, artifact := range artifacts {
-		output.Default.Fprintf(out, " - %s -> ", artifact.ImageName)
-		fmt.Fprintln(out, artifact.Tag)
+		for _, artifact := range artifacts {
+			output.Default.Fprintf(out, " - %s -> ", artifact.ImageName)
+			fmt.Fprintln(out, artifact.Tag)
+		}
 	}
 
 	var localImages []graph.Artifact
@@ -87,15 +89,7 @@ See https://skaffold.dev/docs/pipeline-stages/taggers/#how-tagging-works`)
 	ctx, endTrace := instrumentation.StartTrace(ctx, "Verify_Verifying")
 	defer endTrace()
 
-	// we only want to register images that are local AND were built by this runner OR forced to load via flag
-	var localAndBuiltImages []graph.Artifact
-	for _, image := range localImages {
-		if r.runCtx.ForceLoadImages() || r.wasBuilt(image.Tag) {
-			localAndBuiltImages = append(localAndBuiltImages, image)
-		}
-	}
-
-	r.verifier.RegisterLocalImages(localAndBuiltImages)
+	r.verifier.RegisterLocalImages(localImages)
 	err = r.verifier.Verify(ctx, deployOut, artifacts)
 	postDeployFn()
 	if err != nil {